fix: Port generate_bootstrap_theme.js to sass-embedded
The sass-embedded switch (ba989bc6b5) deleted esbuild/sass_options.js, which
this script still required - website theme compilation was broken on this
branch. It also used the legacy sass.render API that sass-embedded doesn't
support.

Extract the compiler setup (sass-embedded with pure-JS sass fallback, tilde
importer, shared options) from style-plugin.js into esbuild/sass_compiler.js
and use it from both consumers. The theme script now uses compileStringAsync
with sourceMap: false - only the CSS is written, a map would be dead weight.

Error contract with website_theme.py is unchanged: anything on stderr is
treated as failure, success prints only the output path.
